requirements.txt
# ==========================
# Core Deep Learning Stack
# ==========================
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1

# ==========================
# Supporting Libraries
# ==========================
numpy==1.26.4  # ✅ Compatible with all
pandas==2.3.0
scikit-learn==1.3.2
matplotlib==3.8.4
tqdm==4.67.1
Pillow==10.3.0
opencv-python==4.9.0.80
seaborn==0.13.2

# ==========================
# FastAPI + Backend
# ==========================
fastapi==0.110.2
uvicorn==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1

# ==========================
# Visualization & Logging
# ==========================
jsonschema==4.24.0

# ==========================
# AI Validation (CLIP)
# ==========================
open-clip-torch==2.20.0

# ==========================
# Api key integration for species
# ==========================
python-dotenv==1.0.1

# ==========================
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118